+++ packages/expo-clipboard/src/Clipboard.ts
@@ -14,10 +14,6 @@ const onClipboardEventName = 'onClipboardChanged';
 
 type ClipboardEvent = {
   /**
-   * @deprecated Returns empty string. Use [`getStringAsync()`](#getstringasyncoptions) instead to retrieve clipboard content.
-   */
-  content: string;
-  /**
    * An array of content types that are available on the clipboard.
    */
   contentTypes: ContentType[];
@@ -220,20 +216,7 @@ export async function hasImageAsync(): Promise<boolean> {
  * ```
  */
 export function addClipboardListener(listener: (event: ClipboardEvent) => void): EventSubscription {
-  // TODO: Get rid of this wrapper once we remove deprecated `content` property (not before SDK47)
-  const listenerWrapper = (event: ClipboardEvent) => {
-    const wrappedEvent: ClipboardEvent = {
-      ...event,
-      get content(): string {
-        console.warn(
-          "The 'content' property of the clipboard event is deprecated. Use 'getStringAsync()' instead to get clipboard content"
-        );
-        return '';
-      },
-    };
-    listener(wrappedEvent);
-  };
-  return ExpoClipboard.addListener(onClipboardEventName, listenerWrapper);
+  return ExpoClipboard.addListener(onClipboardEventName, listener);
 }
 
 /**